update grouping for sorting; do not coerce ref to char
My initial implementation was incorrect because I was sorting purely
output_type and output_type_id, but Evan brought up the fact that it was
possible for two targets to have separate sets of output type IDs, which
would end up being sorted incorrectly.

I had initially avoided creating a lookup table with the full column
because of the type errors I was getting when trying to compare a date
column with a character column. This is solved by setting `all_character
= FALSE` in `expand_model_out_grid()` and coercing both output_type_id
columns to character.