removed modification of target in test
After the previous commit, this particular test began to fail because
the check started to pass.

During a coworking session with @elray1, we found that the reason why
the check was suddenly passing was because the target "wk ahead inc
covid hosp" was not a valid target according to the hub, thus
`order_output_type_ids()` returned an empty table. The target's presence
was not adding anything important when testing for ascending values with
the output type. Because of this, we determined that this particular
aspect of the test was not necessary to testing the validity of this
function.

I have added a TODO in case this was intentional.